Move sync code.
