+++ crates/cli/src/commands/mod.rs
@@ -14,6 +14,7 @@ pub mod query;
 pub mod run;
 pub mod setup;
 pub mod sync;
+pub mod syncs;
 pub mod task;
 pub mod teardown;
 pub mod upgrade;
