+++ crates/cli/src/commands/syncs/mod.rs
@@ -1,0 +1,1 @@
+pub mod codeowners;
