+++ crates/core/actions/src/lib.rs
@@ -1,0 +1,3 @@
+mod sync_codeowners;
+
+pub use sync_codeowners::*;
